Hyc 1506 indexing service override (#866)
* HYC-1506 adding test and override file

* HYC-1506 add to test

* HYC-1506 delete indexing file and make adjustments

* HYC-1506 fix up indexing_service spec

* HYC-1506 fixing rubocop offenses
